src/requirements.txt
numpy
sounddevice
piper-tts